requirements.txt
crewai>=0.28.0
litellm
langchain-groq
streamlit>=1.32.0
python-dotenv